added online stores, fixed searches and forms
