+++ src/components/HomePage/HomePage.js
@@ -5,6 +5,8 @@ import SearchInput from "./SearchInput";
 import SearchResults from "./SearchResults/SearchResults";
 
 const selectProducts = (reduxState) => {
+  console.log("reduxState", reduxState);
+
   return reduxState.products.single;
 };
 const selectDataNotFound = (reduxState) => {
