+++ src/components/HomePage/SearchResults/Locations/Locations.js
@@ -5,7 +5,7 @@ import AddLocation from "./AddLocationForm/AddLocation";
 export default function Locations({ product, dataNotFound }) {
   return (
     <div>
-      {product && (
+      {product && product.Location && product.Location.length > 0 && (
         <div className="container">
           <Map product={product} />
           {/* <LocationsList stores={product.Store} /> */}
