+++ src/store/city/actions.js
@@ -5,8 +5,6 @@ const baseUrl = "http://localhost:4000";
 export const findCities = (key) => (dispatch) => {
   request(`${baseUrl}/city/${key}`)
     .then((res) => {
-      console.log(res.body);
-
       dispatch(citiesFound(res.body));
     })
     .catch(console.error);
@@ -15,11 +13,5 @@ function citiesFound(cities) {
   return {
     type: "FOUND_CITIES",
     cities,
-  };
-}
-
-export function renewPage() {
-  return {
-    type: "RENEW",
   };
 }
