+++ src/store/online_store/actions.js
@@ -3,8 +3,6 @@ import request from "superagent";
 const baseUrl = "http://localhost:4000";
 
 export const addOnlineStore = (link, productId, countryId) => (dispatch) => {
-  console.log("actiooooons", link, productId, countryId);
-
   request
     .post(`${baseUrl}/online_store`)
     // .set("Authorization", `Bearer ${token}`)
@@ -20,31 +18,3 @@ function storeAdded(newOnlineStore) {
     newOnlineStore,
   };
 }
-
-// export const fetchStore = (id) => (dispatch) => {
-//   request(`${baseUrl}/store/${id}`)
-//     .then((res) => {
-//       dispatch(storeFetched(res.body));
-//     })
-//     .catch(console.error);
-// };
-// function storeFetched(store) {
-//   return {
-//     type: "ONE_STORE",
-//     store,
-//   };
-// }
-
-// export const findStore = (keyword) => (dispatch) => {
-//   request(`${baseUrl}/store/find/${keyword}`)
-//     .then((res) => {
-//       dispatch(storeFound(res.body));
-//     })
-//     .catch(console.error);
-// };
-// function storeFound(stores) {
-//   return {
-//     type: "FIND_STORES",
-//     stores,
-//   };
-// }
